requirements.txt
discord.py
pony